# src/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AutoPro API",
    "description": "Documentacao das rotas da API AutoPro baseada na modelagem Cliente, Veiculo, Servico, Item de Servico e Produto.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
...
  "paths": {
...
    "/perfis": {
      "get": {
        "summary": "Lista todos os perfis",
        "tags": [
          "Perfis"
        ],
        "responses": {
          "200": {
            "description": "Lista de perfis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Perfil"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um perfil",
        "tags": [
          "Perfis"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilInput"
              },
              "example": {
                "nome": "Atendente"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Perfil criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Perfil"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/perfis/{id}": {
      "get": {
        "summary": "Busca um perfil por ID",
        "tags": [
          "Perfis"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numerico do recurso"
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Perfil"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Atualiza um perfil",
        "tags": [
          "Perfis"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numerico do recurso"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilUpdateInput"
              },
              "example": {
                "nome": "Supervisor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Perfil"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "summary": "Remove um perfil",
        "tags": [
          "Perfis"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numerico do recurso"
          }
        ],
        "responses": {
          "200": {
            "description": "Perfil removido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/perfis/{id}/permissoes": {
      "get": {
        "summary": "Lista as permissoes de um perfil",
        "tags": [
          "Perfis"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numerico do recurso"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de permissoes do perfil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permissao"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Substitui as permissoes de um perfil",
        "tags": [
          "Perfis"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numerico do recurso"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilPermissoesInput"
              },
              "example": {
                "ids_permissoes": [
                  1,
                  2,
                  3
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permissoes do perfil atualizadas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Perfil"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/permissoes": {
      "get": {
        "summary": "Lista todas as permissoes",
        "tags": [
          "Permissoes"
        ],
        "responses": {
          "200": {
            "description": "Lista de permissoes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permissao"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria uma permissao",
        "tags": [
          "Permissoes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissaoInput"
              },
              "example": {
                "chave": "usuarios.criar",
                "descricao": "Permite cadastrar usuarios"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Permissao criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permissao"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/permissoes/{id}": {
      "get": {
        "summary": "Busca uma permissao por ID",
        "tags": [
          "Permissoes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numerico do recurso"
          }
        ],
        "responses": {
          "200": {
            "description": "Permissao encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permissao"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Atualiza uma permissao",
        "tags": [
          "Permissoes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numerico do recurso"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissaoUpdateInput"
              },
              "example": {
                "chave": "usuarios.editar",
                "descricao": "Permite editar usuarios"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permissao atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permissao"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "summary": "Remove uma permissao",
        "tags": [
          "Permissoes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numerico do recurso"
          }
        ],
        "responses": {
          "200": {
            "description": "Permissao removida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "Dados invalidos ou obrigatorios ausentes",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "NotFound": {
        "description": "Registro nao encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "Conflict": {
        "description": "Operacao bloqueada por relacionamento ou regra de negocio",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ApiMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Mensagem da API"
          }
        }
      },
...
      "Perfil": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Administrador"
          },
          "permissoes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permissao"
            }
          }
        }
      },
      "PerfilInput": {
        "type": "object",
        "required": [
          "nome"
        ],
        "properties": {
          "nome": {
            "type": "string",
            "example": "Atendente"
          }
        }
      },
      "PerfilUpdateInput": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Supervisor"
          }
        }
      },
      "PerfilPermissoesInput": {
        "type": "object",
        "required": [
          "ids_permissoes"
        ],
        "properties": {
          "ids_permissoes": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "example": [
              1,
              2,
              3
            ]
          }
        }
      },
      "Permissao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "chave": {
            "type": "string",
            "example": "usuarios.criar"
          },
          "descricao": {
            "type": "string",
            "nullable": true,
            "example": "Permite cadastrar usuarios"
          }
        }
      },
      "PermissaoInput": {
        "type": "object",
        "required": [
          "chave"
        ],
        "properties": {
          "chave": {
            "type": "string",
            "example": "usuarios.criar"
          },
          "descricao": {
            "type": "string",
            "nullable": true,
            "example": "Permite cadastrar usuarios"
          }
        }
      },
      "PermissaoUpdateInput": {
        "type": "object",
        "properties": {
          "chave": {
            "type": "string",
            "example": "usuarios.editar"
          },
          "descricao": {
            "type": "string",
            "nullable": true,
            "example": "Permite editar usuarios"
          }
        }
      }
    }
  }
}